Restart rrr-client daily.
BARBIE pointed out an issue where rrr_client misses arbitrary uploads.
He and ANDK have not been able to pinpoint the problem.  It's rare, but
a restart of the rrr-client causes it to pick up missed uploads.  It's
possible that this is the source of some of the issues we have with
dists which are a) not indexed and b) index cleanly once we trigger a
manual indexing.